대환/도커/requirements.txt
torch
torchvision
boto3
Pillow
numpy
segment_anything
flask
